+++ flutter-hms-location/android/src/main/java/com/huawei/hms/flutter/location/constants/Error.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2023. Huawei Technologies Co., Ltd. All rights reserved.
+ * Copyright 2020-2024. Huawei Technologies Co., Ltd. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License.
@@ -19,7 +19,11 @@ package com.huawei.hms.flutter.location.constants;
 public enum Error {
     LOCATION_SETTINGS_NOT_AVAILABLE("Unable to get location settings"),
     NON_EXISTING_REQUEST_ID("Request ID does not exists"),
-    SEND_INTENT_EXCEPTION("Unable to send intent");
+    SEND_INTENT_EXCEPTION("Unable to send intent"),
+    FUSED_LOCATION_NOT_INITIALIZED("FusedLocationService is not initialized."),
+    ACTIVITY_IDENTIFICATION_NOT_INITIALIZED("ActivityIdentificationService is not initialized."),
+    GEOFENCE_SERVICE_NOT_INITIALIZED("GeofenceService is not initialized."),
+    GEOCODER_SERVICE_NOT_INITIALIZED("GeocoderService is not initialized.");
 
     private final String message;