+++ flutter-hms-location/example/lib/screens/activity_identification_screen.dart
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2023. Huawei Technologies Co., Ltd. All rights reserved.
+ * Copyright 2020-2024. Huawei Technologies Co., Ltd. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License.
@@ -58,6 +58,7 @@ class _ActivityIdentificationScreenState
   @override
   void initState() {
     super.initState();
+    _initFusedLocationService();
     _streamSubscription =
         _service.onActivityIdentification.listen(_onIdentificationResponse);
   }
@@ -147,6 +148,15 @@ class _ActivityIdentificationScreenState
     setState(() {
       _bottomText = '$_bottomText\n\n$text';
     });
+  }
+
+  void _initFusedLocationService() async {
+    try {
+      await _service.initActivityIdentificationService();
+      _setTopText('Activity Identification Service has been initialized.');
+    } on PlatformException catch (e) {
+      _setBottomText(e.toString());
+    }
   }
 
   void _createActivityIdentificationUpdates() async {