+++ flutter-hms-location/lib/src/activity/activity_identification_service.dart
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2023. Huawei Technologies Co., Ltd. All rights reserved.
+ * Copyright 2020-2024. Huawei Technologies Co., Ltd. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License.
@@ -54,6 +54,12 @@ class ActivityIdentificationService {
     this._activityIdentificationEventChannel,
     this._activityConversionEventChannel,
   );
+
+  /// Initializes the `Activity Identification` service.
+  Future<void> initActivityIdentificationService() async {
+    await _methodChannel
+        .invokeMethod<void>('initActivityIdentificationService');
+  }
 
   /// Registers for activity identification updates.
   Future<int> createActivityIdentificationUpdates(