+++ flutter-hms-location/lib/src/geofence/geofence_service.dart
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2023. Huawei Technologies Co., Ltd. All rights reserved.
+ * Copyright 2020-2024. Huawei Technologies Co., Ltd. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License")
  * you may not use this file except in compliance with the License.
@@ -43,6 +43,11 @@ class GeofenceService {
     this._methodChannel,
     this._eventChannel,
   );
+
+  /// Initializes the `Geofence` Service.
+  Future<void> initGeofenceService() async {
+    await _methodChannel.invokeMethod<void>('initGeofenceService');
+  }
 
   /// Adds geofences.
   ///